requirements.txt
# Azure Functions
azure-functions>=1.17.0

# HTTP client
aiohttp>=3.9.0
httpx>=0.28.0

# Environment variables
python-dotenv>=1.0.0

# JSON handling
orjson>=3.9.0
